+++ react/src/components/atoms/forms/InputNumber/index.js
@@ -6,13 +6,9 @@ import Input from '../Input';
 import Error from '../Input/error';
 import { InputContext } from '../Input/context';
 import { validNumber } from '../Input/validate';
+import { countDecimals } from '../Input/utility';
 import { numberCharacterPropTypeCheck } from '../../../utilities/componentPropTypeCheck';
 import './style.css';
-
-Number.prototype.countDecimals = function () {
-  if (Math.floor(this.valueOf()) === this.valueOf()) return 0;
-  return this.toString().split('.')[1].length || 0;
-};
 
 const NumberInput = (props) => (
   <React.Fragment>
@@ -33,7 +29,7 @@ const NumberInput = (props) => (
             'ma__input-number-unit--showButtons': props.showButtons
           });
 
-          const decimalPlaces = props.step.countDecimals();
+          const decimalPlaces = countDecimals(props.step);
 
           const displayErrorMessage = (val, min, max, isRequired) => {
             if (isRequired && String(val).length === 0) {
@@ -52,6 +48,14 @@ const NumberInput = (props) => (
               showError: false,
               errorMsg: ''
             };
+          };
+
+          const handleOnBlur = (e) => {
+            const { value } = e.target;
+            const floatValue = Number.parseFloat(value).toFixed(decimalPlaces);
+            if (typeof props.onBlur === 'function') {
+              props.onBlur(e, floatValue);
+            }
           };
 
           const handleChange = (e) => {
@@ -88,6 +92,7 @@ const NumberInput = (props) => (
             maxLength: Number(props.maxlength),
             style: props.width ? { width: `${props.width}px` } : null,
             onChange: handleChange,
+            onBlur: handleOnBlur,
             required: props.required,
             disabled: props.disabled,
             step: props.step
@@ -135,7 +140,7 @@ const NumberInput = (props) => (
 
 const InputNumber = (props) => {
   const {
-    max, min, step, name, onChange, placeholder, width, maxlength, showButtons, ...inputProps
+    max, min, step, name, onChange, onBlur, placeholder, width, maxlength, showButtons, ...inputProps
   } = props;
   // Input and Number share the props.required, props.id and props.disabled values.
   const numberProps = {
@@ -149,6 +154,7 @@ const InputNumber = (props) => {
     required: props.required,
     id: props.id,
     onChange,
+    onBlur,
     disabled: props.disabled,
     unit: props.unit,
     showButtons
@@ -189,6 +195,8 @@ InputNumber.propTypes = {
   errorMsg: PropTypes.string,
   /** Custom change function */
   onChange: PropTypes.func,
+  /** Custom onBlur function */
+  onBlur: PropTypes.func,
   /** Default input value */
   defaultValue: PropTypes.number,
   /** Max value for the field. */
@@ -202,7 +210,7 @@ InputNumber.propTypes = {
   /** A unit that is a string of no more than 2 characters renders in the input after the value, e.g. %  */
   unit: (props, propName) => numberCharacterPropTypeCheck(props, propName, 2),
   /** Whether to render up/down buttons */
-  hasButtons: PropTypes.bool
+  showButtons: PropTypes.bool
 };
 
 InputNumber.defaultProps = {
@@ -210,7 +218,8 @@ InputNumber.defaultProps = {
   required: false,
   onChange: null,
   step: 1,
-  hasButtons: true
+  hasButtons: true,
+  unit: ''
 };
 
 export default InputNumber;